requirements.txt
streamlit
fastapi
uvicorn[standard]
plotly
altair
vega_datasets
pandas
polars
pyarrow
scikit-learn
umap-learn
openTSNE
scanpy
anndata
seaborn
matplotlib
pydantic[dotenv]
python-multipart
joblib
nbformat
jinja2
sqlalchemy
psycopg2-binary
boto3
redis
rq
pre-commit
ruff
pytest
pytest-cov
